scan-parser/requirements.txt
pyyaml
opsgenie-sdk
